move logging config code to /configuration/logging.py
