+++ DEVELOPMENT.md
@@ -67,7 +67,7 @@ Add the following lines to the Python code that invokes BentoML:
 import bentoml
 import logging
 bentoml.config().set('core', 'debug', 'true')
-bentoml.utils.log.configure_logging(logging.DEBUG)
+bentoml.configure_logging(logging.DEBUG)
 ```
 
 And/or use the `--verbose` option when running `bentoml` CLI command, e.g.:
